Lift ThemeProvider to storybook preview, add withTheme class examples (WIP)
