+++ src/shared-components/toggle/index.js
@@ -1,11 +1,9 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 import ToggleButton from 'react-toggle-button';
-import { ThemeProvider } from 'emotion-theming';
-import { primaryTheme, secondaryTheme } from 'src/constants/themes';
+import { useTheme } from 'emotion-theming';
 
-import { COLORS } from '../../constants';
-import { Container, Label, trackStyle, thumbStyle } from './style';
+import { Container, Label, trackStyle, getThumbStyle } from './style';
 
 const propTypes = {
   checked: PropTypes.bool,
@@ -19,47 +17,35 @@ const defaultProps = {
 };
 
 const Toggle = ({ checked, label, onChange }) => {
-  const [theme, setTheme] = React.useState(primaryTheme);
-
-  const toggleTheme = () =>
-    theme.__type === 'primary'
-      ? setTheme(secondaryTheme)
-      : setTheme(primaryTheme);
+  const theme = useTheme();
 
   return (
-    <ThemeProvider theme={theme}>
-      <div>
-        <button onClick={toggleTheme} type="button">
-          swap theme
-        </button>
-      </div>
-      <Container>
-        {label && <Label onClick={onChange}>{label}</Label>}
-        <ToggleButton
-          value={checked || false}
-          onToggle={onChange}
-          inactiveLabel=""
-          activeLabel=""
-          thumbStyle={thumbStyle}
-          trackStyle={trackStyle}
-          thumbAnimateRange={[1, 17]}
-          colors={{
-            active: {
-              base: theme.COLORS.secondary,
-            },
-            inactive: {
-              base: theme.COLORS.primaryTint3,
-            },
-            activeThumb: {
-              base: theme.COLORS.white,
-            },
-            inactiveThumb: {
-              base: theme.COLORS.white,
-            },
-          }}
-        />
-      </Container>
-    </ThemeProvider>
+    <Container>
+      {label && <Label onClick={onChange}>{label}</Label>}
+      <ToggleButton
+        value={checked || false}
+        onToggle={onChange}
+        inactiveLabel=""
+        activeLabel=""
+        thumbStyle={getThumbStyle(theme)}
+        trackStyle={trackStyle}
+        thumbAnimateRange={[1, 17]}
+        colors={{
+          active: {
+            base: theme.COLORS.secondary,
+          },
+          inactive: {
+            base: theme.COLORS.primaryTint3,
+          },
+          activeThumb: {
+            base: theme.COLORS.white,
+          },
+          inactiveThumb: {
+            base: theme.COLORS.white,
+          },
+        }}
+      />
+    </Container>
   );
 };
